Split code into a number of diff helper files, and added common base code for diffing sets
