+++ src/server/helpers/diffFormatters/publisherSet.js
@@ -1,0 +1,56 @@
+/*
+ * Copyright (C) 2016  Ben Ockmore
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.	See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+'use strict';
+
+const set = require('./set');
+
+function formatNewPublisherSet(change) {
+	function transformer(rhs) {
+		return rhs.publishers.map((publisher) => publisher.bbid);
+	}
+
+	return set.formatNewSet(
+		change, 'Publisher', 'publishers', transformer
+	);
+}
+
+function formatPublisherAddOrDelete(change) {
+	return set.formatItemAddOrDelete(
+		change, 'Publisher', (side) => side && [side.bbid]
+	);
+}
+
+function formatPublisherModified(change) {
+	return set.formatItemModified(change, 'Publisher', ['bbid']);
+}
+
+function formatPublisherSet(change) {
+	return set.format(
+		change, 'publisherSet', 'publishers',
+		formatNewPublisherSet,
+		formatPublisherAddOrDelete,
+		formatPublisherModified
+	);
+}
+module.exports.format = formatPublisherSet;
+
+function publisherSetChanged(change) {
+	return set.changed(change, 'publisherSet', 'publishers');
+}
+module.exports.changed = publisherSetChanged;
